refactor: add updateProposaValidationStrategy helper
